SHLVL but before norminette

+++ minishell.h
@@ -539,7 +539,7 @@ int			ft_strncmp(const char *s1, const char *s2, size_t n);
 void		*ft_calloc(size_t nmemb, size_t size, t_pipex *data);
 void		*ft_memcpy(void *dest, const void *src, size_t n);
 
-char		*ft_itoa(int n, t_pipex *data);
+char		*ft_itoa(long n, t_pipex *data);
 
 size_t		ft_digit_count(long int n);
